// src/enums/routesEnum.ts
export enum RoutesEnum {
    HOME = '/',
    MAIN = '/main',
    LOGIN = '/login',
    SIGNUP = '/signup',
    TABLE = '/table',
    TABLE_HIDDEN = '/table-hidden',
    CONTAINER_STOCK = '/tech/container-stock',
    USERS = '/tech/users',
    DELIVERY_CHANNELS = '/tech/delivery-channels',
    STORES = '/tech/stores',
    NOT_FOUND = '/404',
}